stop allowing string control_values for ctrl (#4816)
**Context:**
This warning was added eons ago (#2634), but it pre-dates our
deprecation management page so it slipped through the cracks.

**Description of the Change:**
Remove the support for passing `control_values` as a string to
`qml.ctrl`

**Benefits:**
Deprecation cycle is now complete!

**Possible Drawbacks:**
It wasn't on our deprecations page and was therefore never formally
scheduled for removal. But you know.. it's been almost 10 releases.

NOTE: I want to see PennyLaneAI/plugin-test-matrix#57 merged and get a
look at the matrix before we merge this into master. Too many things
breaking at once 😓