`has_matrix` detects if `Operator.matrix` is overridden (#4844)
Currently, `has_matrix` only is automatically `True` if `compute_matrix`
is overridden. If someone overrides `Operator.matrix` but not
`Operator.compute_matrix`, they will have to manually set `has_matrix =
True`.

This behaviour is a historical remnant from when `Operation` overrode
the `Operator.matrix` method to include in-place inversion. As in-place
inversion is no longer a thing, we no longer have any reason to keep
that behaviour.

With this change, it will be easier for people to write their own custom
operators.


This change is motivated by confusion over why an operator with a matrix
wasn't supported by `default.qubit`.