globalphase decomposes to nothing (#4855)
**Context:**
GlobalPhase is a relatively new operator, and some plugins may not be
able to handle it. To be consistent with the pre-GlobalPhase days, I'm
making it decompose to nothing so devices will decompose it and make it
disappear.

**Description of the Change:**
Implement `GlobalPhase.compute_decomposition` so it decomposes to
nothing

**Benefits:**
Plugins won't break if they see it. It does have a matrix defined, but
maybe some plugins need a whitelist of operators.

**Possible Drawbacks:**
Maybe plugins should fail if they see it, and maybe that's what we
wanted?

**Related GitHub Issues:**
Good to merge with/before #4666 
[sc-50428]

---------

Co-authored-by: Tom Bromley <49409390+trbromley@users.noreply.github.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>