Optimize Hamiltonian-vector products (#4839)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      test directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `doc/releases/changelog-dev.md` file,
summarizing the
      change, and including a link back to the PR.

- [x] The PennyLane source code conforms to
      [PEP8 standards](https://www.python.org/dev/peps/pep-0008/).
We check all of our code against [Pylint](https://www.pylint.org/).
      To lint modified files, simply `pip install pylint`, and then
      run `pylint pennylane/path/to/file.py`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
Hamiltonians are central in QChem workflows. Even for small molecules,
the number of terms quickly grows and tasks like computing expectation
values become bottlenecks. This is partly because there is no
specialized Hamiltonian-statevector product routine. One has to pick
between `sum_of_terms_method` (if `backprop` is required) and
`csr_dot_products`. This PR focuses on optimizing `csr_dot_products`.

In essence, `csr_dot_products` computes the sparse representation of
each term with SciPy's `kron` and sums them up. In #4235 , I introduced
three algorithms to accelerate this:

1. One algorithm exploits the sparsity structure of Pauli words to
directly obtain the CSR format data.
2. One algorithm exploits the redundancy in the sparsity structures of
Pauli words in Pauli sentences to accelerate summing up the sparse
matrices.
3. One algorithm exploits the sparsity of groups of Pauli words (the
fact that there is one non-zero entry per row) together with fast
sorting to further accelerate summing up the sparse matrices.

Together, these improved performance by 4-7x depending on the problem.
Note however, that the speed up is not uniform across 1-3. Indeed, 1. is
much faster than the previous implementation, and hence the bottleneck
becomes entirely 2-3.

These changes were integrated in the Pauli module in #4411, #4475 and
#4272, but PennyLane still uses the
[old](https://github.com/PennyLaneAI/pennylane/blob/master/pennylane/devices/qubit/measure.py#L89)
`sparse_matrix` implementation. To give orders of magnitude, let's
consider the Hamiltonian for an H2 molecule in the CC-PVDZ basis, it has
2951 terms (and the circuit 20 qubits). Simply casting the Hamiltonian
to a `PauliSentence` and using `to_mat` bring the time to generate the
sparse representation down from 84.5 sec to 24.1 sec (on my Latitude
laptop). This however, is only for including the first 1000 terms,
because beyond that the Python process is killed due to lack of memory.
This brings about another issue with computing the full sparse
representation, it demands a lot of memory.

**Description of the Change:**
Since 1. above is so fast and 2,3 are bottlenecks (in both time and
memory), this PR introduces `dot` in `PauliSentence` which takes in a
state vector and returns another state vector equal to the matrix
product. It still takes advantage of 2., but uses the fact that Pauli
words have the sparsity structure of a permutation to avoid summing
sparse matrices and directly evaluates the product using vector
addition.

Even for small molecules, `_pauli_sentence` is so slow it becomes a
bottleneck, and hence the PR proposes a new implementation.

**Benefits:**
Fast Hamiltonian-state vector products with low memory footprint. For
the example above, the product takes about 8 sec. (so ~3x speedup) and
the memory footprint is reduced by ~1000x.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>