pennylane style is non-xkcd, pennylane_sketch is squiggly (#4880)
**Context:**
The PennyLane MPL-drawer style is great, but some users would prefer
smoother plots for more professional contexts, eg. publications. We
should continue to offer the xkcd-like style, but perhaps "pennylane"
should be the non-xkcd style. I propose we make this change, and add
"pennylane_sketch" to keep the squiggly, existing version.

**Description of the Change:**
- Rename the "pennylane" style to "pennylane_sketch"
- Introduce the "pennylane" style to be identical to the old style, but
with non-sketch lines

**Benefits:**
A clean "pennylane" style for professional-looking data plotting, plus
the sketch style is still offered.

**Possible Drawbacks:**
Would be breaking for the "pennylane" style. We can do a deprecation
cycle if people prefer that - it's easy enough to add a warning, just
lmk. I'd prefer not to add the warning, because then people will use the
"pennylane" style that we recommend and get a warning in their face!

[sc-50495]