implement map_wires for cond and measurementValue (#4884)
**Context:**
While updating drawing for `qml.cond`, it was revealed that
`Conditional.map_wires` does not work as expected. I figured it would be
easiest to fix this in a follow-up PR, so here it is. A consequence of
this is that the [teleportation tutorial
broke](https://github.com/PennyLaneAI/qml/actions/runs/6992158970/job/19023455341),
but I double-checked locally that it works as expected with this branch.

**Description of the Change:**
Implement map_wires for `MeasurementValue` and `Conditional`

**Benefits:**
`map_wires` exists for these objects, allowing the drawer to work as
expected, and fixing the qml dev builds

**Possible Drawbacks:**
mapping a `Conditional` maps its `meas_val`, and that maps each
measurement in its `measurements` property. This will preserve the `id`,
as it should. But it might be a little strange. Example:

```python
with qml.queuing.AnnotatedQueue() as q:
    m_0 = qml.measure(0)
    qml.cond(m_0, qml.PauliX)(1)

m_0, cond_op = q.queue
mapped_cond = cond_op.map_wires({0: "a", 1: "b"})
[m_a] = mapped_cond.meas_val.measurements
print(m_0.wires, m_a.wires, m_0.id == m_a.id)
# <Wires = [0]> <Wires = ['a']> True
```

**Related GitHub Issues:**
Fixes #4883 [sc-50825]

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>