JAX (non-jit) uses the `JacobianProductCalculator` and device provided vjps (#4878)
[sc-39465]

This PR ports over the jax (non-jit) interface to use the jacobian
product calculator, and creates a logical pathway to bind vjps with jax
instead of jvps. The vjp pathway would then only be used when we want to
bind device vjps to jax.

Note that we cannot yet support device vjps and jax jacobians, as jax
adds a broadcast dimension to `dy` that default qubit is not yet capable
of handling.


PR #4854 has show we are not quite ready to integrate our PyTree
definitions with how we interact with jax. It currently has some
remaining issues:

* Not all operators are actually valid pytrees, see
`ParametrizedEvolution`, `ParametrizedHamiltonian`,
`HardwareHamiltonian`
* Validation done on initialization prevents the operators from being
used to store cotangents
* What starts out as a vanilla numpy array can be turned into a jax
tracer by the boundary

Instead of reverting that PR to still manually strip out and manually
repack trainable parameters, I decided to close that PR and open a fresh
one.

---------

Co-authored-by: Matthew Silverman <matthews@xanadu.ai>
Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>