lazy-load tape and operator batch sizes (#4911)
**Context:**
We compute operator batch size and initialization, and it's oftentimes
unnecessary. We should only compute it when it's requested

**Description of the Change:**
Only compute operator/tape batch size when they are requested.
Some details:
- inlined the batch-size computation for `ScalarSymbolicOp`
- removed the `_check_batching` overrides in controlled and adjoint
because they are not needed (codecov complained). They both inherit from
`SymbolicOp`, and that doesn't call `super().__init__` so it was never
computing it anyway

**Benefits:**
- Will run expensive computations less frequently
- the Tensorflow issue when pre-computing batch sizes has disappeared!

**Possible Drawbacks:**
the `_check_batching` function did some data validation as well.
operators initialized with poorly-shaped data would fail right away, but
now we'll only see the error when the batch size is requested

**Related GitHub Issues:**
[sc-45972]

---------

Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>