make the pauli_rep operator property public (#4915)
**Context:**
The `pauli_rep` property is widely used, and should be made public

**Description of the Change:**
- Introduce the `pauli_rep` read-only property to the `Operator` class
- Change all (read) references to `op._pauli_rep` into `op.pauli_rep`
(the public version)

**Benefits:**
We're committing to having `pauli_rep` be part of the public API, making
it less scary to lean on while moving forward with other improvements
(eg. combining all the hamiltonian expansion transforms)

**Possible Drawbacks:**
We might want to change things related to it again, and it would require
more careful work because it's public. I'm not so worried about this
because it has existed (and been used) for a while now.

**Related GitHub Issues:**
previously attempted in #4017 
[sc-37494]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>