Exp init error if base is not an operator (#3461)
**Context:**
Running `op=exp(2, base_op)` (coefficient and base_op argument order
backwards) creates an Exp where the base operator is a number and the
coefficient is an operator, without raising an error. Subsequently
calling methods on the created operator either raise errors that are
difficult to understand, or return meaningless results.

**Description of the Change:**
Raise a TypeError on initialization if a `base` is passed to `Exp` and
the base is not an `Operator`.

**Benefits:**
Given how exponentials are written, trying to put arguments in the order
coefficient, operator is an easy mistake to make. Raising an error at
initialization avoids confusing error tracebacks or, in certain cases,
returning results in the correct format that do not answer the intended
question.

---------

Co-authored-by: Tom Bromley <49409390+trbromley@users.noreply.github.com>