Automatically test operator validity (#4922)
**Context:**
Now that we have the `assert_valid` function (#4764), we can easily
validate operators meeting our expectations. The trouble is, we'd have
to test every single operator manually. It would be nice to
automatically generate an instance of each operator (with valid
parameters) and use that to validate the class.

**Description of the Change:**
Add `tests/ops/functions/conftest.py`, which generates a parametrization
of all subclasses of `Operator` (specifics described below). This had to
be put in `conftest.py` because pytest-xdist panics otherwise. Also
update `test_assert_valid.py` to auto-generate an instance of each of
these subclasses and assert that instance's validity. There are some
details to mention about what I've done: I use `op_cls.ndim_params` and
`op_cls.num_params` to generate valid parameters to construct the
instance of the class. This doesn't always work well for various
reasons:
- If they are defined as `property` variables, you can't view it at the
class-level, so I just guess
- I use `len(ndim_params)` to guess the number of parameters, resorting
to `num_params` if that fails
- I use `1.`, `[1.]*(2**n)`, or `np.eye(n)` if the `ndim_params` is 0, 1
or 2 respectively (for each parameter). I assume no other values exist,
and that these will be the appropriate magnitudes. Not always true, eg.
`BasisState` has `ndim_params=(1,)` to get an array, but the length is
`n`, not `2**n`

`_SKIP_OP_TYPES` is a mish-mash of things we don't test. I wrote in-line
comments in the dict to clarify what's what here. Anyway, they are:
- arithmetic ops
- templates
- ops that fail for unknown reasons
- ops that fail BUT have explicit test cases added (currently just a
handful so we can have a test available with a parametrization for the
future)

`_ABSTRACT_OR_META_TYPES` is a set of class types that we shouldn't try
to create instances of. In `get_all_classes`, if `c` is an abstract (eg.
`Operator`, although it's not explicitly stated to be abstract bc
python) or meta (eg. `ControlledOp`, for internal construction only)
class, we don't add it to the final list of classes to make instances
of. We do still add their child classes.

There are also some types that I explicitly xfail in the test. These
_succeed_ in being auto-generated, but they _fail_ validation. I've made
this into a dict, so you have to give a reason explaining why it fails
validation if you're putting it here.

**Benefits:**
- We are more confident in our existing operators!
- Any newly-added Operator will automatically get picked up by this
test. Good ones get added easily, bad ones fail this test in CI and the
dev knows how to fix their operator.

**Possible Drawbacks:**
Only thing I'm worried about now is ensuring that adding tests to make
this totally complete is as easy as possible. might be tough because the
configuration is in `conftest.py`, and not the test file itself. I've
added comments to the test itself so hopefully people look at them.

I'll list possible future work. In order of what I'd like to see:
1. Start adding explicit test cases for everything in `_SKIP_OP_TYPES`
(except qutrit ops, see next task)
2. make `assert_valid` work for qutrit ops then remove the xfail

[sc-51653]

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>