Clarify the distinction between pennylane and vanilla numpy (#4920)
**Context:**
We are working toward distinguishing between pennylane-numpy (autograd)
and vanilla numpy within the codebase. A first step would be to always
import the one we actually intend on importing.

**Description of the Change:**
1. Replace all instances of `from pennylane import numpy as np` with
`import numpy as np`
2. Find the files broken by step 1, and undo those changes, but
importing it as `pnp` to clarify that it's on purpose

I also `import pennylane.numpy` at the top of `__init__.py` because if I
don't, a circular dependency happens (it needs to be imported before the
`kernels` module is imported now that it no longer imports
pennylane-numpy, because it still imports something else that imports
it).

**Benefits:**
- We are closer to making some autograd-only parts of the code
compatible with all interfaces
- We don't use autograd when it isn't necessary

**Possible Drawbacks:**
I might have accidentally removed trainability from something that needs
it? I don't think I have because tests are passing, but maybe that's
just an issue in our test coverage. The optimizers were my main concern,
but surely those tests check for trainability and correctness already.

**Note for QChem reviewers**
I changed every file in qchem/ to use vanilla numpy, and tests failed.
By doing some binary-search undoing, I found that we need
pennylane-numpy in `molecule.py` and that's it. I guess it creates the
variables, and trainability is preserved everywhere else. Maybe this is
a gap in testing, but it looks good now. If someone more familiar with
that part of the code could confirm it, that would be much appreciated

---------

Co-authored-by: Utkarsh <utkarshazad98@gmail.com>