Remove type hinting check in transforms (#4942)
**Context:**

The type hinting check is removed for transforms, in order to make the
transform decorator more practical.

**Description of the Change:**

Remove the type hinting check that would fail if the user does not
respect the transform contract.

The signature check between the transform and the expand transform is
kept.

**Benefits:**

More practical for advanced users

**Possible Drawbacks:**

Potentially harder for beginners to create if the contract is not
understood.

---------

Co-authored-by: Josh Izaac <josh146@gmail.com>