Make `qml.simplify` a transform (#4949)
**Context:**
`qml.simplify` can be applied to operations, measurements, tapes,
qnodes, or qfuncs. It seems appropriate for it to be a transform. So I
made it one.

**Description of the Change:**
* Added `_simplify_transform` with a `transform` decorator to handle
transforming tapes, qnodes, and callables. The public `simplify`
function directly handles operators and measurements, and dispatches to
`_simplify_transform` for other valid inputs.

**Benefits:**
`qml.simplify` is consistent with other op functions/transforms.

**Possible Drawbacks:**
When given a tape, `qml.simplify` used to return a tape, but now it
returns a list with one tape and a processing function. Small breaking
change, but this was also a breaking change for the other op transforms,
so I don't think this is a major deal.

**Related GitHub Issues:**

[sc-52122]

---------

Co-authored-by: Romain Moyard <rmoyard@gmail.com>