Jax-jit can use device vjp (#4935)
[sc-50418]

This PR binds device-provided VJPs to the JAX-JIT interface. It also
switches the jax-jit interface to the "Jacobian Product Calculator"
system of defining how to take Jacobians/ JVPs/ VJPs.

Since we need to place a pure-callback around the jacobian calculation,
but we want to calculate results and jacobians at the same time when
feasible, this PR adds the
`JacobianProductCalculator.execute_and_compute_jacobian` method. This
has the benefit over calling `execute_fn` and `compute_jacobian`
separately, as it is always `grad_on_execution=True` when actually
taking derivatives, and always `grad_on_execution=False` when purely
calculating results.

---------

Co-authored-by: David Wierichs <david.wierichs@xanadu.ai>