Feature/implement probability mp.process counts #4943 (#4952)
### Pull request - issue #4943

**Context: This pull request implements the new feature described in
issue #4943 (internal assignment).**

**Description of the Change:**
Implemented the method `process_counts` in the `ProbabilityMP` class to
compute a probability measurement from a counts dictionary.

**Benefits:**
Representation of store sampling information in the form of a dictionary
counts is much more condensed and memory efficient.

**Possible Drawbacks:** 
The new method should reproduce the results of `process_samples`,
implemented in the same class. It has been tested extensively, although
I still need to become an expert in quantum computing and circuits.
The function checks that the provided sampling information is a
dictionary matching the format returned by `CountsMP`.

**Related GitHub Issues:** 
#4943 


----------------------------------------------------------------------------------------------------------------------------------------------------------------------

**Further details:** 

First of all, congratulations on the library and your fantastic work! 

I had the time to explore and play with the code, which is quite
amazing. I hope to be considered worthy and help you maintain, expand,
and improve it.

I am confident you have improvements and interesting suggestions
regarding my pull request. I left some questions for you below as a
comment on this PR.

Since this is an internal assignment (for which I was given a deadline),
I include some information below that might be relevant for your
evaluation (time zone: GMT-5):

- *Start time:* December 15th, 9:07 AM 
- *Marked as ready for review on:* December 18th, 3:53 AM
- *End time:* TBD
- *Deadline:* December 29th, 9:07 AM

I am incredibly grateful for your time, and I hope to learn so much from
all of you!
I wish you all the best.

---------

Co-authored-by: Christina Lee <chrissie.c.l@gmail.com>