Update MCM statistics tests for `qml.counts` (#4965)
Legacy tests for using `qml.counts` with mid-circuit measurements was
failing due to the probability of one of the outcomes being very low,
causing the counts dictionary to not have the expected number of
entries. I changed the parametrized `RX`s to `Hadamard`s to make the
probability distribution more even. With 1000 shots that pretty much
guarantees that the number of items in the dict will be correct.