Minor MPL drawing sizing fix (#4987)
In PR #4957 , I added extra space to the matplotlib figure and axes so
they could accommodate classical wires without looking odd.

But the extra space was always there, whether or not we actually had any
classical wires. This cropped up when comparing figure outputs from the
qml demos.

This just makes the extra space conditional on the existence of
classical wires, so that previous images will continue looking the same.