Update ControlledSequence decomposition (#4995)
Recreation of PR #4955 , because the merge conflicts with the RC branch
were a mess

**Context:**
The current decomposition of ControlledSequence stops at Power
operations (with non-trainable powers). These are rejected by many ML
interfaces.

**Description of the Change:**

- Added `lazy=False` when creating the power operators.
- Also added a `control_wires` property, because it was returning an
empty wires object, and that's confusing. It does mean we have both
`control` and `control_wires` returning the same thing. If we don't want
both, we could have control_wires raise an error saying it doesn't
exist, but it's strange for it to exist and return empty Wires.

**Benefits:**

- `ControlledSequence` can be used with the ML interfaces.
- The people who come after me will not be confused, as I was, when they
use `control_wires` and everything runs but weird things happen.

---------

Co-authored-by: Matthew Silverman <matthews@xanadu.ai>