composite ops auto-upgrade tensors and hamiltonians (#5031)
**Context:**
We can get some spooky-looking Franken-Operators today, and it often
happens by accident. For example:
```pycon
>>> qml.prod(qml.PauliX(0), 3 * qml.PauliZ(1))
PauliX(wires=[0]) @   (3) [Z1]
```
The second operand became a Hamiltonian instead of an SProd! This makes
sense if you think about the order of python execution and how we
implemented things, but users shouldn't have to think about that.
Moreover, calling `simplify()` on that op won't work so well because
things are all mixed up.

**Description of the Change:**
Update `CompositeOp.__init__` to replace any `Tensor` or `Hamiltonian`
operands with `Prod`, `Sum`, or `SProd` operands as required.

**Benefits:**
User expectations are met, Operator simplification works

**Possible Drawbacks:**
I suspect this will bring far more benefit than drawback, but it's
possible that a user is expecting some weird hybrid operator and is
disappointed by this standardization

[sc-53356]