set xfail_strict=True (#5038)
**Context:**
There is an `xfail_strict` pytest option, which makes xpass (passing
tests that are marked as xfail) tests count as failures.

**Description of the Change:**
- Add `xfail_strict=true` to pytest.ini
- Fix all tests that were xpassing. I tried to make minimal changes to
the tests themselves, but I'm open to more extensive test re-working if
we think it's necessary

**Benefits:**
It's nice to have because it tells us if we've fixed old things we
didn't know we've fixed, which should help us keep our testing etc. up
to date

**Possible Drawbacks:**
If a test is decorated as `pytest.mark.xfail`, we seem to get the most
out of this functionality. However, if you add a `pytest.xfail` WITHIN a
test that would have passed afterwards, pytest doesn't know the
difference and calls this a true xfail (when it should be a failing
xpass)

[sc-52051]

---------

Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>