deprecate qml.matrix with wire_order=None (#5039)
**Context:**
Wire-ordering can surprise people because PennyLane gives no promises.
For example:
```pycon
>>> qs = qml.tape.QuantumScript([qml.Identity(1), qml.CNOT([0, 1])])
>>> qml.matrix(qs)
array([[1., 0., 0., 0.],
       [0., 0., 0., 1.],
       [0., 0., 1., 0.],
       [0., 1., 0., 0.]])
```
This is in wire-order `[1, 0]` because that's the order in which they
appeared in the tape. Most users would expect the other way.

**Description of the Change:**
Deprecate the default `wire_order=None` value in `qml.matrix` in certain
cases. The warning is raised if:
- the object is a tape with more than one wire
- the object is a QNode and the device does not provide wires
- the object is a qfunc

**Benefits:**
Users won't be surprised when they get a matrix in an unexpected wire
order

**Possible Drawbacks:**
More deprecations, the user has to enter wire_order even though it's
often safe to assume they meant ascending integers (but not often
enough, so here we are)

**Related GitHub Issues:**
It was discussed in #3131

[sc-51277]