Clarify difference between `QuantumTape` and `QuantumScript` (#5065)
[sc-54022]

**Context:**

We keep the name `QuantumTape` in documentation, but strictly use its
parent `QuantumScript` in PennyLane source code. We use `QuantumScript`
because carrying around an unnecessary `AnnotatedQueue` can cause
difficulties with parallelism and construction via queuing is
substantially more expensive in both memory and time. Queuing can also
lead to unintended side effects of things entering the circuit that
shouldn't, or things not entering the circuit that should.

This disconnect between documentation and source code is not explained
anyone, and a bit confusing. So this adds some clarification to the
module documentation for `qml.tape.`.

**Description of the Change:**

Added module documentation.

**Benefits:**

Hopefully this clears up some confusion.

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>
Co-authored-by: Matthew Silverman <matthews@xanadu.ai>