Add support for batched cotangents with adjoint VJP (#4963)
**Context:**
This PR adds batching support to `adjoint_vjp`, which allows us to
enable `jax.jacobian` with device provided VJPs on `default.qubit`.

**Description of the Change:**
* Add branches to `adjoint_vjp` to support batched cotangents, both for
expvals and state measurement.
* Add logic for handling inhomogeneous cotangent shapes. This happens
when one of the expvals returned by the tape is not used for the final
result, in which case the cotangents will be a mix between numpy arrays
with some constant number of elements, and the scalar value 0.
* Remove VJP execution related functions from `interfaces/jax.py`. Now,
when using `device_vjp=True`, we always execute with
`jax_jit_vjp_execute` from `interfaces/jax_jit.py`. This is done because
we do not convert to numpy parameters when jitting, and converting to
numpy parameters with causes the resulting jacobian computed with
`adjoint_vjp` to be incorrect. The VJP code removed from `jax.py` was
only used when `device_vjp=True`. Otherwise, the JVP code is used. Thank
you @albi3ro for catching this!
* Remove `pytest.skip` and `pytest.xfail` on tests using `jax.jacobian`
with adjoint diff. There were also a few tests that were still being
skipped when non-expval measurements were made with adjoint diff. I also
stopped skipping those tests.
* Added light documentation about the expected shape of the cotangents
depending on what we are trying to compute with them.

**Benefits:**
* `jax.jacobian` works with adjoint.
* JAX is strictly faster with `device_vjp=True`.
* `jax.jacobian` is now vectorized when jitting with `device_vjp=True`.

**Possible Drawbacks:**
None

**Related GitHub Issues:**

---------

Co-authored-by: albi3ro <chrissie.c.l@gmail.com>
Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: Matthew Silverman <matthews@xanadu.ai>
Co-authored-by: lillian542 <38584660+lillian542@users.noreply.github.com>
Co-authored-by: Tom Bromley <49409390+trbromley@users.noreply.github.com>
Co-authored-by: Utkarsh <utkarshazad98@gmail.com>
Co-authored-by: Romain Moyard <rmoyard@gmail.com>
Co-authored-by: Josh Izaac <josh146@gmail.com>