legacy device: check if object is measurement before checking stopping_condition (#5071)
**Context:**
Many plugins are failing because of the deprecation of
`MeasurementProcess.name` (#5047). The old device API checks if an
object meets the stopping requirement _or_ if it's a measurement (in
that order). Due to this ordering, we reference `mp.name` which is now
deprecated.

**Description of the Change:**
Swap the order of conditions in `_local_tape_expand` to avoid hitting
this deprecation warning

**Benefits:**
Plugins will work again!

**Possible Drawbacks:**
N/A