fix `undo_swaps` and `split_non_commuting` (#5081)
When exploring in PR #5058, I found bugs in `undo_swaps` and
`split_non_commuting`.

[sc-54373]
[sc-54374]

Undo swaps had a fragile way of mapping wires, as it was created before
we had `Operator.map_wires`. Now it works with more general operations.

`split_non_commuting` was not passing the shots along. This was a silent
issue on the old device interface, but would be causing problems if it
was ever used with the new device interface.

---------

Co-authored-by: Thomas R. Bromley <49409390+trbromley@users.noreply.github.com>