Update `operation._process_data` to consider abstract parameters (#3917)
**Context:**

When using a device that does not support broadcasting together with a
broadcasted tape, the tape will be decomposed into multiple tapes, one
for each parameter(s) value along the broadcasting axis.
When using this together with JIT (see an example here: #2762) and
caching (which is activated by default), the decomposed tapes will have
the same hash, so that only the first tape will be executed and the
cached result for this tape will be reused for the other tapes. Note
that this bug only occurs when using `diff_method="backprop"`.

**Description of the Change:**
This PR changes the private method `Operator._process_data`, which
returns a string representation of the data of the `Operator`, to
consider abstract parameters: When abstract parameters are in
`Operation.data`, they are included in the string representation via
their `id`. This leads to distinct representations for distinct abstract
values and allows caching to work correctly.

**Benefits:**
Fixes #2762

**Possible Drawbacks:**
N/A

**Related GitHub Issues:**
fixes #2762

---------

Co-authored-by: Christina Lee <christina@xanadu.ai>
Co-authored-by: albi3ro <chrissie.c.l@gmail.com>