restore test for `stop_recording` cleanup (#5094)
This was originally added in #3182, but I only put the test in
`test_tape.py`, so it was accidentally pruned in #3281 (anagram PR!)
which isn't ideal. I'm putting it back in `test_queuing.py` where it
truly belongs.

I know this PR seems a bit insignificant, but I remember this bug coming
up and I don't want it to resurface because of a lack of testing. Plus,
the `finally` is much prettier than writing the cleanup code twice 😄

Not bothering with a changelog entry, but can add if we feel strongly
about it