re-add support for tf.function with abstract shape (#5089)
**Context:**
While working on adding poetry support, [this qcut test
failed](https://github.com/PennyLaneAI/pennylane/actions/runs/7613788036/job/20734550376)
in a familiar way - it tried to check batching with an abstract
TensorShape (of None)! Sub-context: it hasn't been run for a _very_ long
time because it needs networkx<3 and poetry is making me use that,
unlike today's CI runs. I thought I had fixed this in #4911 mostly
because tests stopped failing. However, they (eg. the test that I am
changing here) stopped failing because we were no longer _running_ the
batch-size evaluation. It's a niche corner case (see that PR description
for more details), but it used to pass and it now fails on master.

**Description of the Change:**
- Re-add the previous functionality to `Operator._check_batching()`
where it skips the check if the operator params are abstract TF tensors
with a shape of None in its signature within a `tf.function`
- Removed the `MyRX` from the test because it only made sense when I
added it (in other words, before I re-defined `Operator.ndim_params` in
the linked PR), but am now calling `op.batch_size` to ensure the loading
of the property

**Benefits:**
Tests that used to pass are passing again, specifically if you use
`tf.function` with `input_signature=(tf.TensorSpec(shape=None))`

**Possible Drawbacks:**
Additional try-except logic (re-)added to `Operator._check_batching()`
which is called frequently